question_generation/pyproject.toml
[project]
name = "question-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.68.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
